improve parsing-files: structure = create_map.c and create_texture.c

+++ libft/include/libft.h
@@ -44,7 +44,7 @@ void	*ft_memmove(void *dest, const void *src, size_t n);
 char	*ft_strnstr(const char *big, const char *little, size_t n);
 size_t	ft_strlcpy(char *dst, const char *src, size_t size);
 size_t	ft_strlcat(char *dst, const char *src, size_t size);
-int		ft_atoi(const char *nptr);
+size_t	ft_atoi(const char *nptr);
 int		ft_toupper(int c);
 int		ft_tolower(int c);
 void	*ft_calloc(size_t nmemb, size_t size);
